Start a merge of sidplay2fp changes.  This incorporates the changes for PSID/RSIDv3.  It also includes some additional checks since sidplay2fp would also support some of these things incorrectly in PSID/RSIDv2 (both reads and writes).
Other changes have been included that are easy cleanups, but there are still more changes in SidTune.cpp that need reviewing.  Other changes related to header path changes, using stdint.h directly, merging sidplay2, libsidplay2, libsidutils or dropping support for older file formats cannot be accepted.

VC2008 even now does not support stdint.h (I do not have 2010 to tell if thats better).  Projects are separate to reduce size and because the are used separately by other projects.  The header structure mirrors that of Linux else it is problematic to build code against the un-installed AND INSTALLED versions.

There were/are plans to break out sidtune since it is in use by a number of utility projects:

-hvscupdate
-various checking utils
-libsidplay1

Dropping of support for older gcc and VC5/6 is a consideration.  Not sure about the others since I only have limited access to compilers newer than 2003.